tests/wpt/tests/html/webappapis/dynamic-markup-insertion/document-write/023.html

<!doctype html>
<title>document.write</title>
<script src="/resources/testharness.js"></script><script src="/resources/testharnessreport.js"></script>
<script>
var t = async_test();
t.step(function() {
  document.write("<body><span ");
});
</script>id=a>Filler Text</span>
<script>
t.step(function() {
  assert_equals(document.body.childNodes[0].localName, "span");
  assert_equals(document.body.childNodes[0].getAttribute("id"), "a");
  assert_equals(document.body.childNodes[0].textContent, "Filler Text");
}
);
t.done();
</script>
<div id="log"></div>
